refactor(ai-prompt-decorator): migrate ai-prompt-decorator to new framework
