+++ spec/03-plugins/41-ai-prompt-decorator/02-integration_spec.lua
@@ -3,7 +3,7 @@ local helpers = require "spec.helpers"
 local PLUGIN_NAME = "ai-prompt-decorator"
 
 
-for _, strategy in helpers.all_strategies() do if strategy ~= "cassandra" then
+for _, strategy in helpers.all_strategies() do
   describe(PLUGIN_NAME .. ": (access) [#" .. strategy .. "]", function()
     local client
 
@@ -88,7 +88,7 @@ for _, strategy in helpers.all_strategies() do if strategy ~= "cassandra" then
 
       assert.response(r).has.status(400)
       local json = assert.response(r).has.jsonbody()
-      assert.same(json, { error = { message = "this LLM route only supports llm/chat type requests" }})
+      assert.same({ error = { message = "this LLM route only supports llm/chat type requests" }}, json)
     end)
 
 
@@ -107,9 +107,9 @@ for _, strategy in helpers.all_strategies() do if strategy ~= "cassandra" then
 
       assert.response(r).has.status(400)
       local json = assert.response(r).has.jsonbody()
-      assert.same(json, { error = { message = "this LLM route only supports llm/chat type requests" }})
+      assert.same({ error = { message = "this LLM route only supports llm/chat type requests" }}, json)
     end)
 
   end)
 
-end end
+end
